jpashop/src/main/java/jpabook/jpashop/MemberRepository.java
package jpabook.jpashop;

import org.springframework.stereotype.Repository;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

@Repository
public class MemberRepository {

    @PersistenceContext
    private EntityManager em;

    // member를 왜 반환하지 않을까?
    // command와 query를 분리하기 위함.
    // save는 side effect 가능성이 있는 command성 이기 때문
    public Long save(Member member){
        em.persist(member);
        return member.getId();
    }

    public Member find(Long id){
        return em.find(Member.class, id);
    }
}
